feat(v0.7.3): pre-mic-button hardening — 4 work streams
Work Stream 1 — Bug Fixes:
- A-04: useAudioRecorder — add useEffect unmount cleanup to prevent OS
  mic-indicator leak when component navigates away mid-recording
- A-05: useSpeechToText — add cancelledRef guard; cancel() sets the ref
  so any in-flight transcribeAudio/formatTranscript awaits short-circuit
  instead of calling setFormattedResult on stale state
- A-08: RichTextEditor — change L2/L3 insertContent → tr.insertText for
  formattedResult.formatted and .raw; eliminates HTML injection surface
  from Ollama/OpenAI responses (L1 insertContent left unchanged)
- A-10: useSpeechToText — wire isAvailable to availabilityResultRef
  instead of stale settings.modelDownloaded flag

Work Stream 2 — CI Local Alias:
- Add check:deny, check:audit, check:all scripts to package.json
- Add Makefile with check/deny/audit/typecheck targets

Work Stream 3 — Tests (+56 tests, 450 total):
- New: src/hooks/useSpeechToText.test.ts (7 tests) — model guard,
  cancel race × 2, quickCapture bypass, L2/L3 overlay path, L2/L3
  local fallback, isAvailable from ref (A-10)
- Expand: src/lib/aiService.test.ts (+4 tests) — L2 Ollama happy path,
  L3 OpenAI happy path, L2 AbortError fallback, L3 401 fallback

Work Stream 4 — SetupScreen Decomposition:
- Extract 10 step components into src/components/setup/ (WelcomeStep,
  SourceStep, PasswordStep, RecoveryStep, SecurityStep, StorageStep,
  DevicesStep, SyncFromPeerStep, ImportStep, CompleteStep)
- SetupScreen reduced from 1365 LOC to ~230 LOC orchestrator; all
  shared state stays in orchestrator, steps own only ephemeral UI state
